ANDROID: dma-heap: Rework allocation calls to return struct dma_buf instead of fd
While most uses will want to allocate a fd for a dmabuf, there
are some cases where it might be useful to have just a dma_buf structure.

So refactor the heap allocaiton functions to return a struct dma_buf
and add a dma_heap_bufferfd_alloc() function to convert that return a fd.

[CPNOTE: 06/07/21] Lee: Asked vendor for an update via the bug

Signed-off-by: John Stultz <john.stultz@linaro.org>
Bug: 154341375

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: Id6fd7e8471b9555a08bcdf8ca210feb589fa51c1
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>