BACKPORT: FROMLIST: media: v4l2-core: extend the v4l2 format to support request
This patch is to extend the related interface to support the request-based
operations. We added request fd to the parameters of MEDIA_IOC_SETUP_LINK,
VIDIOC_S_FMT, VIDIOC_SUBDEV_S_SELECTION, VIDIOC_SUBDEV_S_FMT.
The driver uses media_request_get_by_fd() to retrieve the media request and
save the pending change in it, so that we can apply the pending change in
req_queue() callback then.

Signed-off-by: Louis Kuo <louis.kuo@mediatek.com>

Fix the spacing problem in the comment
(am from https://patchwork.kernel.org/patch/12239311/)

Change-Id: I291268e7981e7b2f224db076ca14e2d35cd3a265
Signed-off-by: Chihhao Tsai <chihhao.tsai@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>