arm64: dts: mediatek: mt8365-evk: set pull up on HDMI irq line
This commit fixes an interrupt storm by setting a pull up on the HDMI
irq line. The IRQ line is active low so it needs to stay high during
normal operations, and since there is no pull-up on the board, we need
to set the pull-up in the input pin.

Signed-off-by: Fabien Parent <fparent@baylibre.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>