STOPSHIP: drm: mediatek: mtk_dp: Fix eDP+DP dual-port BDG detect resume issues
The BDG detect function is unconditionally returning
connector_status_connected for both DP and eDP: while for
the eDP case this is ok (as the bridge won't be attached
if there's no panel in the first place), for the DP case
this is not, as this relies on plug-state detection.

By always returning connector_status_connected on the DP
interface, if no DP display is really connected, the DRM
will try to enable the bridge, resulting in a timeout and
recovery attempts, which will also disable the eDP panel,
resulting in .. nothing working anymore.

Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>