BL_INTERNAL: HACK: soc: mediatek: mt8195: make audio's power domain always-on
DMIC hangs during suspend-to-ram (S3) if audio is recorded. The cause is
still being investigated.

To avoid the hang, keep the audio's power domain always-on, this will
prevent power/context loss on the audio device.

This is a pure HACK and a proper fix should be introduced to fix DMIC
and no longer use always-on flag.

Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>