FIXUP: GENIO: drm/mediatek: hdmi: Fix hpd status parsing
This fixes the issue that some monitors not waking up
when connected to Genio 1200 demo board during start up,
but works normally if we remove and re-plug the HDMI connector.

The root cause is that HDMI driver does not correctly detect
HPD (Hot Plug Detection) status, even if HPD pin is
always be pulled high after plugged in.

According to status register definition, HPD_PIN_STA BIT(4) indicates
HPD status, so it should be returned whenever plugged in.

fixes: e2d208d858b8 (BL_INTERNAL: drm/mediatek: add mtk8195 hdmi display driver)
Change-Id: Ieb15c2e567a5e4b2f91a8eb679a0a60ef04afe36
CR-Id: AUTO00203855
Signed-off-by: mac shen <mac.shen@mediatek.com>
Signed-off-by: Tommy Chen <tommyyl.chen@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>