FROMLIST: media: platform: mtk-mdp3: chip config split about component settings
For different chips, different types and numbers of components are
configured in the MDP to achieve the desired application, which will
cause ID compatibility problems in the same driver.

Subdivide the component ID into two kinds:
	1. public_id (compatible with other chips)
	2. inner_id (what the current chip actually owns)

In addition, the chip configuration related structures of the component
are integrated and moved into the header file.

Link: https://patchwork.kernel.org/project/linux-mediatek/patch/20221124123859.24395-3-moudy.ho@mediatek.com/

Change-Id: Ifaaf05ad67a34721000d7710aee00a64d5addf5b
Signed-off-by: Moudy Ho <moudy.ho@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>