GENIO: FIXME: clk: mediatek: mt8188: Probe vpp1 with simple_probe
The current design of clk/pm_genpd would fail if any one of the
clks in device tree is not successfully probed.

The actual clock managers of vdo0/1 and vpp0/1 are implemented
in soc/mediatek mmsys modules.

Use simple_probe until the corresponding mmsys drivers are ready,
so driver developers could use nodes such as

/sys/kernel/debug/clk/clk_summary
/sys/kernel/debug/pm_genpd/pm_genpd_summary

Change-Id: I86c06fbda1ae006e7715b413936b8dd723e2c500
Signed-off-by: Chris-QJ Chen <chris-qj.chen@mediatek.com>
Signed-off-by: Pablo Sun <pablo.sun@mediatek.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>