GENIO: mtk-vcu-driver: mt8395: Move vcud_task to local
The task structure, vcud_task, is used to store the process
who opens the vcu. It was maintained as a global pointer.
That is, each vcu instance may overwrite it. Therefore, move
this pointer to the instance scope. It should be independent
of others.

Change-Id: If8f1991ac03eff68067d2c3eb8809f2d2e7f63bb
Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>
(backported from https://gitlab.com/mediatek/aiot/bsp/mtk-vcu-driver/-/commit/a5cf85327cfe4a8195391869455316d14eb79231)
Signed-off-by: Jian Hui Lee <jianhui.lee@canonical.com>
Signed-off-by: Hsuan-Yu Lin <shane.lin@canonical.com>