GENIO: mtk-vcu-driver: mt8395: Create device at the end of probe
This patch is to fix the timing issue by relocating device creation
to the end of the probe function. This ensures that all necessary
preparations are completed before exposing the interface to the user,
preventing the issue of a partially initialized driver.

In addition, this patch also makes the error handling symmetrical by

1. Create vcu_ipi_deinit() which corresponds to vcu_ipi_init().
2. Allocate vcu after obtaining vcuid.
3. Go to vcu_mutex_destroy when gce_cmds allocation fails.

Signed-off-by: Andy.Hsieh <andy.hsieh@mediatek.com>
Change-Id: I7e1e7edc7d555853d2a31df1512a01b689061d2e